Negative random value where positive value is required
The function MythRandom(), introduced in commit 31a811e54e21ac4160a3ed533f965eb9caeb4f7b,
does create a 32-bit random value which is returned in an uint32_t.
This new function does replace the Linux system call random() which returns an int.
Note that int value returned by random() is always positive.
In places where an int is required the output of MythRandom() is cast to an int.
This value can however be negative and will be approximately 50% of the times it is called.
This is now fixed in the places where it had not been fixed alreay.

+++ mythtv/programs/mythfrontend/playbackbox.cpp
@@ -2254,7 +2254,7 @@ void PlaybackBox::playSelectedPlaylist(bool Random)
         QList<uint> tmp = m_playList;
         while (!tmp.isEmpty())
         {
-            uint i = static_cast<int>(MythRandom()) % tmp.size();
+            uint i = MythRandom() % tmp.size();
             m_playListPlay.append(tmp[i]);
             tmp.removeAll(tmp[i]);
         }
